SO-1957: close index if there is one

+++ commons/com.b2international.index.api.tests.tools/src/com/b2international/index/revision/BaseRevisionIndexTest.java
@@ -73,7 +73,9 @@ public abstract class BaseRevisionIndexTest {
 
 	@After
 	public void teardown() {
-		index.admin().delete();
+		if (index != null) {
+			index.admin().delete();
+		}
 	}
 	
 	protected void configureMapper(ObjectMapper mapper) {
